#CLOUD-328 one more unit test
#VMS-2270 fixed: 'find system' multicast packet did not contain cloudSystemId. That resulted in very long cloud system discovery in LAN
#VMS-2391 CloudServerSocket sends listen request after connection to mediator has been restored
#VMS-2391 RetryTimer: added reset to inital state for re-usage
#VMS-2391 added declaration of subscription to stun client tcp reconnect event
#VMS-2391 fixed compile error
#vms-2354. Low quality archive playback restored.
#vms-2401. Archive camera info file is created on per storage basis rather than per storage pool.
#vms-2401. Archive camera info file is created on per storage basis rather than per storage pool. Merge heads.
#vms. Merge prod_3.0.0 -> dev_3.0.0.
CLOUD-324 network::cloud::AddressResolver fix, tests and refactoring
VMS-2154 Put all NICs up before calculating HardwareID on Windows
address review commments
fixed compile error
fixed compile warnings
minor compile fix
some review fixes

<==== Mercurial data ====>
Branch: dev_3.0.0_gui

+++ common_libs/nx_network/src/nx/network/stun/unreliable_message_pipeline.h
@@ -308,7 +308,7 @@ private:
         }
         else
         {
-            NX_ASSERT(bytesSent == m_sendQueue.front().serializedMessage.size());
+            NX_ASSERT(bytesSent == (size_t)m_sendQueue.front().serializedMessage.size());
         }
 
         auto completionHandler = std::move(m_sendQueue.front().completionHandler);
